Make Chezmoi check also DISPLAY env variables for headless check
This commit makes Chezmoi detect if `DISPLAY` or `WAYLAND_DISPLAY`
variables are set. This suggests that the system is NOT headless.

This is particular useful as Chezmoi might be run inside a `distrobox`
container. Using previous detection it would detect that the system is
headless even though its not.

This fixes it. Also this should increase compatibilty with Wayland
desktops